requirements.txt
python-dotenv==0.21.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
gunicorn==20.1.0
Django==4.1.3
django-filter==22.1
django-import-export==3.0.1
